requirements.txt
asgiref==3.8.1
black==24.4.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
identify==2.5.36
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pre-commit==3.7.0
PyJWT==2.8.0
PyYAML==6.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
virtualenv==20.26.1
